api/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
openai
serper
python-multipart
